Process the SINR correctly
There was a bug that was causing an incorrect SINR processing.

Fixed and added the option to chose the TX Power and the Number of available channels.

Channels are still being chosen randomly. Next step: planification.

Next step regarding SINR: process SINR map to detect the best areas and the worst ones.